Complete 3rd javascript exercise
This exercise requires us to reverse the string.
My first attempt uses a for loop to access
individual elements from the end and copy them
until it reaches to the first element. This took
4 lines of code.

However the solution condenses the method into
1 line w/ return statement where you split the
word in order to reverse it, then join it back
into the string.